Create Unit-3 Program-2
Write a NumPy program using NumPy methods - max, min, argmax, argmin, argmax, repr, count, bincount, unique.  
a. To extract all numbers from a given array which are less and greater than a specified number.  
b. To find the indices of the maximum and minimum numbers along the given axis of an array.